+++ hooks/emitter/decoder.go
@@ -1,18 +1,16 @@
 package emitter
 
 import (
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	"encoding/hex"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	transfertypes "github.com/cosmos/cosmos-sdk/x/ibc/applications/transfer/types"
-	clienttypes "github.com/cosmos/cosmos-sdk/x/ibc/core/02-client/types"
-	types1 "github.com/cosmos/cosmos-sdk/x/ibc/core/02-client/types"
-	connectiontypes "github.com/cosmos/cosmos-sdk/x/ibc/core/03-connection/types"
-	channeltypes "github.com/cosmos/cosmos-sdk/x/ibc/core/04-channel/types"
-	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
+
+	"github.com/bandprotocol/chain/v2/hooks/common"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	clienttypes "github.com/cosmos/ibc-go/modules/core/02-client/types"
+	connectiontypes "github.com/cosmos/ibc-go/modules/core/03-connection/types"
+	channeltypes "github.com/cosmos/ibc-go/modules/core/04-channel/types"
 
 	"github.com/bandprotocol/chain/v2/hooks/common"
 	oracletypes "github.com/bandprotocol/chain/v2/x/oracle/types"
@@ -211,6 +209,63 @@ func decodeMsgCreateClient(msg *clienttypes.MsgCreateClient, detail common.JsDic
 	detail["client_state"] = clientState
 	detail["consensus_state"] = consensusState
 	detail["signer"] = msg.Signer
+}
+
+func decodeMsgSubmitProposal(msg *govtypes.MsgSubmitProposal, detail common.JsDict) {
+	detail["content"] = msg.GetContent()
+	detail["initial_deposit"] = msg.GetInitialDeposit()
+	detail["proposer"] = msg.GetProposer()
+}
+
+func decodeMsgDeposit(msg *govtypes.MsgDeposit, detail common.JsDict) {
+	detail["proposal_id"] = msg.ProposalId
+	detail["depositor"] = msg.Depositor
+	detail["amount"] = msg.Amount
+}
+
+func decodeMsgVote(msg *govtypes.MsgVote, detail common.JsDict) {
+	detail["proposal_id"] = msg.ProposalId
+	detail["voter"] = msg.Voter
+	detail["option"] = msg.Option
+}
+
+func decodeMsgCreateValidator(msg *stakingtypes.MsgCreateValidator, detail common.JsDict) {
+	pk, _ := msg.Pubkey.GetCachedValue().(cryptotypes.PubKey)
+	hexConsPubKey := hex.EncodeToString(pk.Bytes())
+
+	detail["description"] = msg.Description
+	detail["commission_rates"] = msg.Commission.Rate
+	detail["min_self_delegation"] = msg.MinSelfDelegation
+	detail["delegator_address"] = msg.DelegatorAddress
+	detail["validator_address"] = msg.ValidatorAddress
+	detail["pubkey"] = hexConsPubKey
+	detail["value"] = msg.Value
+}
+
+func decodeMsgEditValidator(msg *stakingtypes.MsgEditValidator, detail common.JsDict) {
+	detail["description"] = msg.Description
+	detail["validator_address"] = msg.ValidatorAddress
+	detail["commission_rates"] = msg.CommissionRate
+	detail["min_self_delegation"] = msg.MinSelfDelegation
+}
+
+func decodeMsgDelegate(msg *stakingtypes.MsgDelegate, detail common.JsDict) {
+	detail["delegator_address"] = msg.DelegatorAddress
+	detail["validator_address"] = msg.ValidatorAddress
+	detail["amount"] = msg.Amount
+}
+
+func decodeMsgUndelegate(msg *stakingtypes.MsgUndelegate, detail common.JsDict) {
+	detail["delegator_address"] = msg.DelegatorAddress
+	detail["validator_address"] = msg.ValidatorAddress
+	detail["amount"] = msg.Amount
+}
+
+func decodeMsgBeginRedelegate(msg *stakingtypes.MsgBeginRedelegate, detail common.JsDict) {
+	detail["delegator_address"] = msg.DelegatorAddress
+	detail["validator_src_address"] = msg.ValidatorSrcAddress
+	detail["validator_dst_address"] = msg.ValidatorDstAddress
+	detail["amount"] = msg.Amount
 }
 
 func decodeMsgUpdateClient(msg *clienttypes.MsgUpdateClient, detail common.JsDict) {