+++ x/oracle/genesis.go
@@ -47,7 +47,7 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, data *types.GenesisState) {
 	// Only try to bind to port if it is not already bound, since we may already own
 	// port capability from capability InitGenesis
 	if !k.IsBound(ctx, types.PortID) {
-		// transfer module binds to the transfer port on InitChain
+		// oracle module binds to the oracle port on InitChain
 		// and claims the returned capability
 		err := k.BindPort(ctx, types.PortID)
 		if err != nil {